backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
firebase-admin==6.2.0
Pillow==10.1.0
rembg==2.0.50
requests==2.31.0
python-multipart==0.0.6
psutil==5.9.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
unidecode==1.3.7
openai==1.3.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
pillow-heif==0.15.0
ftfy==6.1.3
git+https://github.com/openai/CLIP.git 